displays files with clickable download button, still working on backend for it to send the file downloads, added logout button, changed home page to be "home" for all users
